wx/requirements.txt
requests>=2.31
python-dotenv>=1.0
markdown>=3.5
